переделал получение кол-ва мест и кол-ва свободных мест

+++ flight-service/src/main/java/ru/kalinin/flight/dto/response/FlightAdminResponse.java
@@ -18,7 +18,7 @@ public class FlightAdminResponse {
     private LocalDateTime departureTime;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime arrivalTime;
-    private Integer totalSeats;
-    private Integer availableSeats;
+    private Long totalSeats;
+    private Long availableSeats;
     private List<SeatResponse> seats;
 }
